Extract evaluation and export from worker
