+++ bindings/wasm/examples/lib/manifold-gltf.ts
