+++ bindings/wasm/examples/make-manifold.ts
@@ -18,7 +18,7 @@ import {prune} from '@gltf-transform/functions';
 import {SimpleDropzone} from 'simple-dropzone';
 
 import Module from './built/manifold.js';
-import {disposeMesh, Properties, readMesh, setupIO, writeMesh} from './gltf-io';
+import {disposeMesh, Properties, readMesh, setupIO, writeMesh} from './lib/gltf-io';
 
 // Set up gltf-transform
 const io = setupIO(new WebIO());
