+++ bindings/wasm/examples/tsconfig.json
@@ -18,7 +18,9 @@
     "skipLibCheck": true,
   },
   "include": [
-    "./*.ts"
+    "./*.ts",
+    "lib/*.ts"
   ],
   "exclude": []
 }
+
